pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-gpt"
version = "0.1.0"
description = "Offline voice transcription, journaling, and memory retrieval"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "voice-gpt"}]
dependencies = [
  "numpy>=1.24",
  "faiss-cpu>=1.7.4",
  "sentence-transformers>=2.6.1",
]

[project.optional-dependencies]
cli = ["typer>=0.12", "rich>=13.7"]
faster = ["faster-whisper>=1.0.3"]

[project.scripts]
voice-gpt = "voice_gpt.cli:app"

[tool.setuptools]
packages = ["voice_gpt"]
